Split database config and global config
