+++ core/database/config.json
@@ -1,0 +1,3 @@
+{
+    "fileName": "discotron.db"
+}
